JS-Library-independent Version
The Line 'moo_onepagewebsitenavigation'	=>
'system/modules/onepagewebsite/templates', in autoload.php has no
effect. clicked or unchecked in Layout the Script will all-time be
included.
Also it have to be MooTools.
With this commit the line in autoload.php will have an effect and there
will be a choice for jQuery with the same features. Also to use with
the plugin »smoothscroll« from jQuery